refactor(tigrbl_auth): move RFC endpoints into rfc package
