+++ pkgs/standards/tigrbl_auth/tigrbl_auth/rfc/rfc6749_token.py
@@ -13,31 +13,36 @@ from sqlalchemy import select
 from tigrbl.engine import HybridSession as AsyncSession
 from pydantic import ValidationError
 
-from ...backends import AuthError
-from ...fastapi_deps import get_db
-from ...orm import AuthCode, Client, DeviceCode, User
-from ...rfc8707 import extract_resource
-from ...runtime_cfg import settings
-from ...rfc6749 import (
+from ..backends import AuthError
+from ..fastapi_deps import get_db
+from ..orm import AuthCode, Client, DeviceCode, User
+from .rfc8707 import extract_resource
+from ..runtime_cfg import settings
+from .rfc6749 import (
     RFC6749Error,
     enforce_authorization_code_grant,
     enforce_grant_type,
     enforce_password_grant,
     is_enabled as rfc6749_enabled,
 )
-from ...rfc7636_pkce import verify_code_challenge
-from ...rfc8628 import DeviceGrantForm
-from ...oidc_id_token import mint_id_token, oidc_hash
-from ...rfc8414_metadata import ISSUER
+from .rfc7636_pkce import verify_code_challenge
+from .rfc8628 import DeviceGrantForm
+from ..oidc_id_token import mint_id_token, oidc_hash
+from .rfc8414_metadata import ISSUER
 
-from ..schemas import (
+from ..routers.schemas import (
     AuthorizationCodeGrantForm,
     PasswordGrantForm,
     RefreshIn,
     TokenPair,
 )
-from ..shared import _require_tls, _jwt, _pwd_backend, _ALLOWED_GRANT_TYPES
-from . import router
+from ..routers.shared import (
+    _require_tls,
+    _jwt,
+    _pwd_backend,
+    _ALLOWED_GRANT_TYPES,
+)
+from ..routers.authz import router
 
 
 @router.post("/token", response_model=TokenPair)
